+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+
+## 2.6.0
 * Add `columnSettings` prop, that enables column order and visibility settings dialog
 * Add `defaultSortColumn` and `defaultSortOrder` to `grid` props that enables column default sorting